package/requirements.txt
python-bitcoinlib
requests
